initial implementation of SUBB. it has problems.
there is something wrong with either the glue/chain between SUBB and its
carry flag input, or something else is going on. An expression with two
different subtractions will crash the compiler because it deletes the
carry-clearing between them (it actually tries to save the 0 carry-in on
the stack and restore it. i have a hack "slack" carry flag so it doesn't
actually generate a spill, but this still doesn't work)